TEH: updated Task 1, added the code for Task 2
Templated and Exception Handling:
-Task 1 now meets the specified requirements
-Task 2 complete

+++ TEMPLATES AND EXCEPTION HANDLING/Task 1/main.cpp
@@ -9,8 +9,8 @@ template <typename T> T Sum(T x, T y) {
 
 int main()
 {
-    float a, b;
-    double c, d;
+    int a, b;
+    float c, d;
 
     cin >> a >> b >> c >> d;
